+++ explorer/client/src/components/ownedobjects/OwnedObjectResultType.tsx
@@ -1,0 +1,14 @@
+// Copyright (c) 2022, Mysten Labs, Inc.
+// SPDX-License-Identifier: Apache-2.0
+
+import type BN from 'bn.js';
+
+export type DataType = {
+    id: string;
+    Type: string;
+    _isCoin: boolean;
+    Version?: string;
+    display?: string;
+    balance?: BN;
+}[];
+
