+++ explorer/client/src/components/ownedobjects/OwnedObjects.tsx
@@ -38,17 +38,6 @@ import Longtext from '../longtext/Longtext';
 import Pagination from '../pagination/Pagination';
 import PaginationLogic from '../pagination/PaginationLogic';
 
-import styles from './OwnedObjects.module.css';
-
-type resultType = {
-    id: string;
-    Type: string;
-    _isCoin: boolean;
-    Version?: string;
-    display?: string;
-    balance?: BN;
-    name?: string;
-}[];
 
 const DATATYPE_DEFAULT: resultType = [
     {
@@ -182,52 +171,7 @@ function OwnedObjectAPI({ id, byAddress }: { id: string; byAddress: boolean }) {
 
 const viewFn = (results: any) => <OwnedObjectView results={results} />;
 
-function OwnedObjectLayout({ results }: { results: resultType }) {
-    const coin_results = results.filter(({ _isCoin }) => _isCoin);
-    const other_results = results
-        .filter(({ _isCoin }) => !_isCoin)
-        .sort((a, b) => {
-            if (a.Type > b.Type) return 1;
-            if (a.Type < b.Type) return -1;
-            if (a.Type === b.Type) {
-                return a.id <= b.id ? -1 : 1;
-            }
-            return 0;
-        });
 
-    const nftFooter = {
-        stats: {
-            count: other_results.length,
-            stats_text: 'Total NFTs',
-        },
-    };
-
-    return (
-        <div className={styles.layout}>
-            {coin_results.length > 0 && (
-                <div>
-                    <div className={styles.ownedobjectheader}>
-                        <h2>Coins</h2>
-                    </div>
-                    <GroupView results={coin_results} />
-                </div>
-            )}
-            {other_results.length > 0 && (
-                <div id="NFTSection">
-                    <div className={styles.ownedobjectheader}>
-                        <h2>NFTs</h2>
-                    </div>
-                    <PaginationLogic
-                        results={other_results}
-                        viewComponentFn={viewFn}
-                        itemsPerPage={ITEMS_PER_PAGE}
-                        stats={nftFooter.stats}
-                    />
-                </div>
-            )}
-        </div>
-    );
-}
 
 function GroupView({ results }: { results: resultType }) {
     const CLOSED_TYPE_STRING = '';
@@ -359,115 +303,6 @@ function GroupView({ results }: { results: resultType }) {
                 onPagiChangeFn={setCurrentPage}
             />
         </>
-    );
-}
-
-function OwnedObjectView({ results }: { results: resultType }) {
-    return (
-        <div id="ownedObjects" className={styles.ownedobjects}>
-            {results.map((entryObj, index1) => (
-                <div className={styles.objectbox} key={`object-${index1}`}>
-                    {entryObj.display !== undefined && (
-                        <div className={styles.previewimage}>
-                            <DisplayBox
-                                display={entryObj.display}
-                                caption={
-                                    entryObj.name ||
-                                    trimStdLibPrefix(entryObj.Type)
-                                }
-                            />
-                        </div>
-                    )}
-                    <div className={styles.textitem}>
-                        {Object.entries(entryObj).map(
-                            ([key, value], index2) => (
-                                <div key={`object-${index1}-${index2}`}>
-                                    {(() => {
-                                        switch (key) {
-                                            case 'Type':
-                                                if (entryObj._isCoin) {
-                                                    break;
-                                                } else {
-                                                    return (
-                                                        <span
-                                                            className={
-                                                                styles.typevalue
-                                                            }
-                                                        >
-                                                            {trimStdLibPrefix(
-                                                                value as string
-                                                            )}
-                                                        </span>
-                                                    );
-                                                }
-                                            case 'balance':
-                                                if (!entryObj._isCoin) {
-                                                    break;
-                                                } else {
-                                                    return (
-                                                        <div
-                                                            className={
-                                                                styles.coinfield
-                                                            }
-                                                        >
-                                                            <div>Balance</div>
-                                                            <div>
-                                                                {String(value)}
-                                                            </div>
-                                                        </div>
-                                                    );
-                                                }
-                                            case 'id':
-                                                if (entryObj._isCoin) {
-                                                    return (
-                                                        <div
-                                                            className={
-                                                                styles.coinfield
-                                                            }
-                                                        >
-                                                            <div>Object ID</div>
-                                                            <div>
-                                                                <Longtext
-                                                                    text={String(
-                                                                        value
-                                                                    )}
-                                                                    category="objects"
-                                                                    isCopyButton={
-                                                                        false
-                                                                    }
-                                                                    alttext={alttextgen(
-                                                                        value
-                                                                    )}
-                                                                />
-                                                            </div>
-                                                        </div>
-                                                    );
-                                                } else {
-                                                    return (
-                                                        <Longtext
-                                                            text={String(value)}
-                                                            category="objects"
-                                                            isCopyButton={false}
-                                                            alttext={alttextgen(
-                                                                value
-                                                            )}
-                                                        />
-                                                    );
-                                                }
-                                            default:
-                                                break;
-                                        }
-                                    })()}
-                                </div>
-                            )
-                        )}
-                    </div>
-                </div>
-            ))}
-            {lastRowHas2Elements(results) && (
-                <div className={`${styles.objectbox} ${styles.fillerbox}`} />
-            )}
-        </div>
     );
 }
