+++ explorer/client/src/components/ownedobjects/views/OwnedObjects.module.css
